# openapi.yaml
openapi: 3.0.0
info:
  title: Fuse
  version: 1.0.4
servers:
- description: sandbox
  url: https://sandbox-api.letsfuse.com
security:
- fuseApiKey: []
- fuseClientId: []
...
paths:
  /v1/financial_connections/accounts/details:
    post:
      operationId: getFinancialConnectionsAccountDetails
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetFinancialConnectionsAccountDetailsRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetFinancialConnectionsAccountDetailsResponse'
          description: Success
      summary: Get account details
      tags:
      - Fuse
    summary: Get account details
  /v1/financial_connections/owners:
    post:
      operationId: getFinancialConnectionsOwners
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetFinancialConnectionsOwnersRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetFinancialConnectionsOwnersResponse'
          description: Success
      summary: Get account owners
      tags:
      - Fuse
  /v1/financial_connections/accounts:
    post:
      operationId: getFinancialConnectionsAccounts
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetFinancialConnectionsAccountsRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetFinancialConnectionsAccountsResponse'
          description: Successful response
      summary: Get accounts
      tags:
      - Fuse
  /v1/asset_report/create:
    post:
      description: Use this endpoint to generate an Asset Report for a user.
      operationId: createAssetReport
      requestBody:
        content:
          application/json:
            examples:
              "0":
                value: "{\n\t\"access_token\": \"3e0cc775-3588-460d-863a-a4baee8ac7c5\"\
                  ,\n\t\"days_requested\": 90,\n}"
            schema:
              $ref: '#/components/schemas/CreateAssetReportRequest'
      responses:
        "200":
          content:
            application/json:
              examples:
                "0":
                  value: "{\"asset_report_token\":\"fuse_ar_dev_1MV8sa\",\"asset_report_id\"\
                    :\"3256bf0f-6881-4ff4-9f27-54c396572320\"}"
              schema:
                $ref: '#/components/schemas/CreateAssetReportResponse'
          description: Response
      tags:
      - Fuse
    summary: Create an Asset Report
  /v1/asset_report/refresh:
    post:
      description: Refreshes the Asset Report in JSON format.
      operationId: refreshAssetReport
      requestBody:
        content:
          application/json:
            examples:
              "0":
                value: "{\n\t\"asset_report_token\": \"fuse_ar_dev_1MV8sa\"\n}"
            schema:
              $ref: '#/components/schemas/RefreshAssetReportRequest'
      responses:
        "200":
          content:
            application/json:
              examples:
                "0":
                  value: "{\"asset_report_token\":\"fuse_ar_dev_1MV8sa\",\"asset_report_id\"\
                    :\"3256bf0f-6881-4ff4-9f27-54c396572320\"}"
              schema:
                $ref: '#/components/schemas/CreateAssetReportResponse'
          description: Response
      tags:
      - Fuse
    summary: Refresh Asset Report
  /v1/asset_report:
    post:
      description: Retrieves the Asset Report in JSON format.
      operationId: getAssetReport
      requestBody:
        content:
          application/json:
            examples:
              "0":
                value: "{\n\t\"asset_report_token\": \"fuse_ar_dev_1MV8sa\"\n}"
            schema:
              $ref: '#/components/schemas/GetAssetReportRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAssetReportResponse'
          description: Response
      tags:
      - Fuse
    summary: Get an asset report
  /v1/financial_connections/balances:
    post:
      operationId: getFinancialConnectionsBalances
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetFinancialConnectionsBalanceRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetFinancialConnectionsBalanceResponse'
          description: Successful response
      summary: Get balances
      tags:
      - Fuse
  /v1/entities/{entity_id}:
    get:
      operationId: getEntity
      parameters:
      - explode: false
        in: path
        name: entity_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetEntityResponse'
          description: Success
      summary: Get an entity - an entity is automatically created after a successful
        institution connection
      tags:
      - Fuse
    summary: Get entity
  /v1/financial_connections/public_token/exchange:
    post:
      description: API to exchange a public token for an access token and financial
        connection id
      operationId: exchangeFinancialConnectionsPublicToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeFinancialConnectionsPublicTokenRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeFinancialConnectionsPublicTokenResponse'
          description: Response
      tags:
      - Fuse
    summary: Exchange public token
  /v1/financial_connections/sync:
    post:
      description: Call this endpoint upon receiving a financial_connection.sync_data
        webhook. This will keep the financial connections data up to date.
      operationId: syncFinancialConnectionsData
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncFinancialConnectionsDataRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncFinancialConnectionsDataResponse'
          description: Successfully synchronized transactions
      summary: Sync financial connections data
      tags:
      - Fuse
  /v1/financial_connections/institutions/{institution_id}:
    get:
      description: Receive metadata for a financial institution
      operationId: getFinancialInstitution
      parameters:
      - explode: false
        in: path
        name: institution_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetFinancialInstitutionResponse'
          description: Financial institution metadata
      summary: Get a financial institution
      tags:
      - Fuse
  /v1/financial_connections/{financial_connection_id_to_delete}:
    delete:
      operationId: deleteFinancialConnection
      parameters:
      - explode: false
        in: path
        name: financial_connection_id_to_delete
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteFinancialConnectionResponse'
          description: Success
      summary: Delete a financial connection
      tags:
      - Fuse
    summary: Delete a financial connection
  /v1/financial_connections/{financial_connection_id}:
    get:
      operationId: getFinancialConnection
      parameters:
      - explode: false
        in: path
        name: financial_connection_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetFinancialConnectionResponse'
          description: Success
      summary: Get financial connection details
      tags:
      - Fuse
    summary: Get financial connection details
  /v1/financial_connections/migrate:
    post:
      description: "This endpoint migrates financial connections from Plaid or MX\
        \ into the unified Fuse API. It accepts a POST request with connection data,\
        \ aggregator, entity, and Fuse products, and responds with a JSON payload\
        \ containing the migrated connection's data, access token, ID, and request\
        \ ID."
      operationId: migrateFinancialConnection
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MigrateFinancialConnectionsTokenRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MigrateFinancialConnectionsTokenResponse'
          description: Success
      summary: Migrate financial connection
      tags:
      - Fuse
    summary: Migrate financial connection
  /v1/financial_connections/investments/holdings:
    post:
      operationId: getInvestmentHoldings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetInvestmentHoldingsRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetInvestmentHoldingsResponse'
          description: Successful response
      summary: Get investment holdings
      tags:
      - Fuse
  /v1/financial_connections/investments/transactions:
    post:
      operationId: getInvestmentTransactions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetInvestmentTransactionsRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetInvestmentTransactionsResponse'
          description: Successful response
      summary: Get investment transactions
      tags:
      - Fuse
  /v1/financial_connections/liabilities:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetLiabilitiesRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetLiabilitiesResponse'
          description: Successful response
      summary: Get liabilities
      tags:
      - Fuse
  /v1/link/token:
    post:
      description: Create a link token to start the process of a user connecting to
        a specific financial institution.
      operationId: createLinkToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLinkTokenRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateLinkTokenResponse'
          description: Response
      tags:
      - Fuse
    summary: Create link token
  /v1/session:
    post:
      description: Creates a session that returns a client_secret which is required
        as a parameter when initializing the Fuse SDK.
      operationId: createSession
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSessionResponse'
          description: Response
      tags:
      - Fuse
    summary: Create session
  /v1/financial_connections/transactions:
    post:
      operationId: getFinancialConnectionsTransactions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetFinancialConnectionsTransactionsRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetFinancialConnectionsTransactionsResponse'
          description: Success
      summary: Get transactions
      tags:
      - Fuse
    summary: Get transactions
components:
  schemas:
...
    GetFinancialConnectionsAccountDetailsRequest:
      example:
        access_token: access_token
      properties:
        access_token:
          description: Access token for authentication
          title: access_token
          type: string
      required:
      - access_token
      title: GetFinancialConnectionsAccountDetailsRequest
      type: object
...
    GetFinancialConnectionsAccountDetailsResponse:
      example:
        financial_connection:
          id: id
          institution_id: institution_id
        account_details:
        - remote_id: remote_id
          ach:
            wire_routing: wire_routing
            routing: routing
            bacs_routing: bacs_routing
            account: account
        - remote_id: remote_id
          ach:
            wire_routing: wire_routing
            routing: routing
            bacs_routing: bacs_routing
            account: account
        request_id: request_id
      properties:
        account_details:
          items:
            $ref: '#/components/schemas/FinancialConnectionsAccountDetails'
          title: account_details
          type: array
        financial_connection:
          $ref: '#/components/schemas/FinancialConnectionData'
        request_id:
          description: An identifier that is exclusive to the request and can serve
            as a means for investigating and resolving issues.
          title: request_id
          type: string
      required:
      - account_details
      - financial_connection
      - request_id
      title: GetFinancialConnectionsAccountDetailsResponse
      type: object
    GetFinancialConnectionsOwnersRequest:
      example:
        access_token: access_token
      properties:
        access_token:
          description: Access token for authentication
          title: access_token
          type: string
      required:
      - access_token
      title: GetFinancialConnectionsOwnersRequest
      type: object
...
    GetFinancialConnectionsAccountsRequest:
      example:
        access_token: access_token
      properties:
        access_token:
          description: Access token for authentication
          title: access_token
          type: string
      required:
      - access_token
      title: GetFinancialConnectionsAccountsRequest
      type: object
    GetFinancialConnectionsAccountsResponse:
      example:
        financial_connection:
          id: id
          institution_id: institution_id
        accounts:
        - institution:
            name: name
          balance:
            current: 6.027456183070403
            available: 0.8008281904610115
            iso_currency_code: iso_currency_code
            last_updated_date: last_updated_date
          subtype: null
          remote_id: remote_id
          fingerprint: fingerprint
          name: name
          type: null
          mask: mask
        - institution:
            name: name
          balance:
            current: 6.027456183070403
            available: 0.8008281904610115
            iso_currency_code: iso_currency_code
            last_updated_date: last_updated_date
          subtype: null
          remote_id: remote_id
          fingerprint: fingerprint
          name: name
          type: null
          mask: mask
        request_id: request_id
      properties:
        accounts:
          items:
            $ref: '#/components/schemas/FinancialConnectionsAccount'
          title: accounts
          type: array
        financial_connection:
          $ref: '#/components/schemas/FinancialConnectionData'
        request_id:
          description: An identifier that is exclusive to the request and can serve
            as a means for investigating and resolving issues.
          title: request_id
          type: string
      required:
      - accounts
      - financial_connection
      - request_id
      title: GetFinancialConnectionsAccountsResponse
      type: object
    CreateAssetReportRequest:
      example:
        access_token: access_token
        include_identity: true
        days_requested: 30.15014613278082
      properties:
        access_token:
          description: Access token for the entity to create report for.
          title: access_token
          type: string
        days_requested:
          description: The maximum integer number of days of history to include in
            the Asset Report
          maximum: 365
          minimum: 1
          title: days_requested
          type: number
        include_identity:
          default: true
          description: Indicates whether to include identity data in the Asset Report
          title: include_identity
          type: boolean
      required:
      - access_token
      - days_requested
      title: CreateAssetReportRequest
      type: object
...
    SyncFinancialConnectionsDataRequest:
      description: Unified webhook data
      type: object
...
  securitySchemes:
    fuseApiKey:
      in: header
      name: Fuse-Api-Key
      type: apiKey
    fuseClientId:
      in: header
      name: Fuse-Client-Id
      type: apiKey
    plaidClientId:
      in: header
      name: Plaid-Client-Id
      type: apiKey
    plaidSecret:
      in: header
      name: Plaid-Secret
      type: apiKey
    tellerApplicationId:
      in: header
      name: Teller-Application-Id
      type: apiKey
    tellerCertificate:
      in: header
      name: Teller-Certificate
      type: apiKey
    tellerPrivateKey:
      in: header
      name: Teller-Private-Key
      type: apiKey
    tellerTokenSigningKey:
      in: header
      name: Teller-Token-Signing-Key
      type: apiKey
    tellerSigningSecret:
      in: header
      name: Teller-Signing-Secret
      type: apiKey
    mxClientId:
      in: header
      name: Mx-Client-Id
      type: apiKey
...